test(infra): extract shared postgres IT base and migrate flyway tests
