+++ src/test/java/com/rubenrzprz/reshub/FlywayReservationsIT.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
 
-class FlywayReservationsIT extends FlywayITBase {
+class FlywayReservationsIT extends FlywayIntegrationTestBase {
 
   @Test
   void v4TablesConstraintsAndIndexesExist() {
