+++ src/test/java/com/rubenrzprz/reshub/support/PostgresIntegrationTestBase.java
@@ -1,0 +1,28 @@
+package com.rubenrzprz.reshub.support;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+public abstract class PostgresIntegrationTestBase {
+
+  @SuppressWarnings("resource")
+  static final PostgreSQLContainer<?> postgres =
+    new PostgreSQLContainer<>("postgres:16-alpine")
+      .withDatabaseName("reshub")
+      .withUsername("reshub")
+      .withPassword("reshub");
+
+  static {
+    postgres.start();
+  }
+
+  @DynamicPropertySource
+  static void register(DynamicPropertyRegistry r) {
+    r.add("spring.datasource.url", postgres::getJdbcUrl);
+    r.add("spring.datasource.username", postgres::getUsername);
+    r.add("spring.datasource.password", postgres::getPassword);
+    r.add("spring.jpa.hibernate.ddl-auto", () -> "validate");
+    r.add("spring.flyway.enabled", () -> "true");
+  }
+}
